+++ libexec/scoop-download.ps1
@@ -18,9 +18,11 @@
 #   -s, --skip-hash-check           Skip hash verification (use with caution!)
 #   -u, --no-update-scoop           Don't update Scoop before downloading if it's outdated
 #   -a, --arch <32bit|64bit|arm64>  Use the specified architecture, if the app supports it
+#   -w, --virustotal-check          Check the download against VirusTotal (may be slow)
 
 . "$PSScriptRoot\..\lib\getopt.ps1"
 . "$PSScriptRoot\..\lib\json.ps1" # 'autoupdate.ps1' (indirectly)
+. "$PSScriptRoot\..\lib\helper\hash.ps1" # 'hash_for_url'
 . "$PSScriptRoot\..\lib\autoupdate.ps1" # 'generate_user_manifest' (indirectly)
 . "$PSScriptRoot\..\lib\versions.ps1" # 'Select-CurrentVersion'
 . "$PSScriptRoot\..\lib\manifest.ps1" # 'generate_user_manifest' 'Get-Manifest'
@@ -30,12 +32,13 @@ if (get_config USE_SQLITE_CACHE) {
     . "$PSScriptRoot\..\lib\database.ps1"
 }
 
-$opt, $apps, $err = getopt $args 'fsua:' 'force', 'skip-hash-check', 'no-update-scoop', 'arch='
+$opt, $apps, $err = getopt $args 'fsua:w' 'force', 'skip-hash-check', 'no-update-scoop', 'arch=', 'virustotal-check'
 if ($err) { error "scoop download: $err"; exit 1 }
 
 $check_hash = !($opt.s -or $opt.'skip-hash-check')
 $use_cache = !($opt.f -or $opt.force)
 $architecture = Get-DefaultArchitecture
+$check_virustotal = $opt.w -or $opt.'virustotal-check' -or (get_config USE_VIRUSTOTAL $false)
 try {
     $architecture = Format-ArchitectureString ($opt.a + $opt.arch)
 } catch {
@@ -101,9 +104,16 @@ foreach ($curr_app in $apps) {
     }
 
     if(Test-Aria2Enabled) {
-        Invoke-CachedAria2Download $app $version $manifest $architecture $cachedir $manifest.cookie $use_cache $curr_check_hash
+        Invoke-CachedAria2Download $app $version $manifest $architecture $cachedir $manifest.cookie $use_cache $curr_check_hash $check_virustotal
     } else {
-        foreach($url in script:url $manifest $architecture) {
+        $urls = @(script:url $manifest $architecture)
+        $urls = if ($check_virustotal) {
+            Test-UrlsWithVirusTotal $app $urls $manifest $architecture
+        } else {
+            $urls
+        }
+
+        foreach($url in $urls) {
             try {
                 Invoke-CachedDownload $app $version $url $null $manifest.cookie $use_cache
             } catch {